Refactor `CreateCommitments` tests suites (#485)
* merge TransferCommitmentsTestSuite and CreateCommitmentsTestSuite

* fix SetupTest in CreateCommitmentsTestSuite

* simplify MassMigrationCommitmentsTestSuite by using testSuiteWithTxsContext

* rename MassMigrationCommitmentsTestSuite to MMCommitmentsTestSuite

* rename Create2TransferCommitmentsTestSuite to C2TCommitmentsTestSuite

* Remove redundant teardown methods
